refactor: split on transport logic from icap client
