+++ lib/Net/ICAP/ResponseParser.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Files_Antivirus\ICAP;
+namespace OCA\Files_Antivirus\Net\ICAP;
 
 use RuntimeException;
